requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2025.1.31
cffi==1.17.1
cryptography==44.0.0
DateTime==5.5
decorator==5.2.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
marzban==0.4.3
MonthDelta==0.9.1
paramiko==3.5.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyNaCl==1.5.0
python-telegram-bot==21.10
pytz==2025.1
schedule==1.2.2
setuptools==75.8.0
sniffio==1.3.1
sshtunnel==0.4.0
typing_extensions==4.12.2
zope.interface==7.2
sqlmodel~=0.0.24